+++ src/find.ts
@@ -1,5 +1,5 @@
 import { showTable } from './presentation.js';
-import { findCycleTime } from './domain/cycle-time/services/find-cycle-time.js';
+import { findCycleTime } from './services/find-cycle-time.js';
 
 const main = async (pullNumber: number) => {
   const res = await findCycleTime(pullNumber);
