+++ src/index.ts
@@ -1,5 +1,5 @@
 import { showTable } from './presentation.js';
-import { pullCycleTimes } from './domain/cycle-time/services/pull-cycle-times.js';
+import { pullCycleTimes } from './services/pull-cycle-times.js';
 
 const main = async (perPage: number, page: number) => {
   const res = await pullCycleTimes(perPage, page);
